# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nestora API - Business platform for small businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── BUSINESS ────────────────────────────────────────────────────────────────
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Category deleted

  # ─── PRODUCTS ────────────────────────────────────────────────────────────────
...
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List active offers
      responses:
        "200":
          description: Offer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create an offer (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}:
    patch:
      operationId: updateOffer
      tags: [offers]
      summary: Update an offer (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    delete:
      operationId: deleteOffer
      tags: [offers]
      summary: Delete an offer (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Offer deleted
  /offers/validate:
    post:
      operationId: validateCoupon
      tags: [offers]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Coupon valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
        "404":
          description: Invalid coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── LOYALTY ─────────────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ─── ANALYTICS ───────────────────────────────────────────────────────────────
...
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer profile (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [owner, customer]
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, loyaltyPoints, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, customer]
        phone:
          type: ["string", "null"]
        loyaltyPoints:
          type: integer
        createdAt:
          type: string

    # BUSINESS
...
    Category:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        imageUrl:
          type: string

    # PRODUCTS
...
    Offer:
      type: object
      required: [id, title, type, discountPercent, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [festival, flash, coupon, loyalty]
        discountPercent:
          type: number
        code:
          type: ["string", "null"]
        minOrderAmount:
          type: ["number", "null"]
        maxUses:
          type: ["integer", "null"]
        usedCount:
          type: integer
        validFrom:
          type: ["string", "null"]
        validTo:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    OfferInput:
      type: object
      required: [title, type, discountPercent]
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [festival, flash, coupon, loyalty]
        discountPercent:
          type: number
          minimum: 0
          maximum: 100
        code:
          type: string
        minOrderAmount:
          type: number
        maxUses:
          type: integer
        validFrom:
          type: string
        validTo:
          type: string
        isActive:
          type: boolean
    OfferUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        discountPercent:
          type: number
        code:
          type: string
        minOrderAmount:
          type: number
        maxUses:
          type: integer
        validFrom:
          type: string
        validTo:
          type: string
        isActive:
          type: boolean
    CouponValidateInput:
      type: object
      required: [code]
      properties:
        code:
          type: string

    # LOYALTY
...
    Notification:
      type: object
      required: [id, userId, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: ["string", "null"]
        message:
          type: string
        type:
          type: string
          enum: [order, offer, loyalty, ticket, general]
        isRead:
          type: boolean
        createdAt:
          type: string

    # ANALYTICS
...
    CustomerProfile:
      type: object
      required: [id, name, email, loyaltyPoints, totalOrders, totalSpent, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        loyaltyPoints:
          type: integer
        totalOrders:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string